Update graphs to handle new data formatting
Specifically, amp-icmp and amp-tcpping data now includes a
'median' field and the percentiles are contained in a field
called 'rtts' (not 'rtt' like it used to be). Loss is also
calculated differently.

Unfortunately, DNS has not changed from 'rtt' to 'rtts' (yet?) so
latency graphs still need to support the old style.

The matrix still needs testing with recent data, but the tooltips
seem to work OK.

Rainbow traceroute needs a complete revision, still, but that's
tomorrow's job.