Started making the traceroute map work.
Updated Brad Christensen's traceroute map code to work with our
new data format.

It currently draws semi-useful graphs but there's still a lot
we want to do like add weight to heavily used paths and nodes,
special symbols for timeouts/errors and colouring based on AS.
I've started attaching useful information to the nodes to aid in
this, but don't make use of it yet.

Fixed bug where Smokeping graphs were not being drawn because the
data formatting code did not like the Decimals that were being
returned by ampy. (Not fixed in the last commit, like I said.)