Traceroute map is now in a working state
Fixed bug where page was using 100% CPU constantly.

Added weight metrics to edges and correctly calculate unique paths
and visits for any highlighted edge.

Fixed node tooltips to show IP address rather than the internal node ID.

Add extra "error" node to the end of paths that resulted in an ICMP error.

Highlight all paths that an edge contributes to when the edge is moused over.

Still lots to do to make the map look and behave the way we want -- will
add a ticket to cover the main tasks.