Fix bad selection queries on the latency modal
If you've added a DNS series and then click the "add new series" button,
the modal would attempt to find ICMP and TCPPing series that match the
current set of DNS selection options. This will cause ampy to report an
error because there are more DNS options than ICMP options.

Instead, the queries for populating each tab will now only consider options
that belong to that tab rather than the options for the currently selected tab.