Fix bad highlighting on traceroute rainbow graph
If an AS appeared twice, horizontally matching bands were not being
recognised properly. Borders would be drawn around each individual measurement
rather than simply once around the whole band. This is now fixed, provided
the band is same height the whole way across.

It's a bit harder to make this work for regions where the y0 and y1 values
vary from measurement to measurement, as we'd have to do multiple draws.