Highlight AS hops by changing fill colour
Previously, we used to draw a border around the highlighted hop, but
this was difficult to do correctly whenever the hop appeared in multiple
disjoint parts of the path. Just tweaking the fill colour slightly is just
as effective at catching the eye and much easier to achieve.

Modified stroke styles to ensure the colour changes are modest rather than
disorienting.