Add AS info to the traceroute map graph
Nodes are now coloured based on AS rather than IP address.

The AS name now appears on the hover tooltip for nodes.

Commented out path_loop in drawHit function for tracemap -- this
doesn't appear to be doing anything and it is not clear what the
original purpose was. I think it is meant to highlight all paths that
pass through the hit node, but it isn't even managing that right now.