Added support for AMP HTTP graphs
This resulted in adding a new graphstyle: "http". This style is
essentially the same as a basic time series, except the tooltip
will show extra information about the page size, object count and
server count.

Added ability to 'encode' values fetched from modal dropdowns. This
is required for HTTP because the target is a URL and we run into
problems if we try to put a URL in our AJAX requests. Instead, if
a dropdown is flagged as requiring encoding, all '/' characters are
replaced with '|' characters and the URL is also encoded.

Replacing the slashes is necessary because pyramid will auto-decode the URL
when parsing the request, so it will still see the extra slashes and
interpret them as such.