Use leftmost X value when returning a hit on a series
Previously, we were returning the closest X value to the mouse
pointer. This meant that the values returned for a horizontal line
would depend on which side of the line you moused over. This was
resulting in some confusing behaviour with the new HTTP graph, where
the download rate would not be consistent with the values reported
in the tooltip.