Allow enabling / disabling of matrix tabs
The 'ampweb.matrixtabs' config option can be used to specify which
tabs should appear on the matrix and the order in which they should appear.

If the option is absent, latency, hops and http tabs are automatically
provided.

The titles, descriptions and IDs for each tab is now controlled by the
appropriate Collection class in the views code, rather than hard-coded into
the template.

Updated dist config files to include the new option.