Don't add two URLs to history when using eventview
eventview was redirecting to a URL based on the collection (e.g.
amp-icmp) which was then being converted to a URL with the graphstyle
(e.g. amp-latency).

This had the effect of adding two entries to the session history when
accessing a graph via the event dashboard, meaning that the back button
would not return the user to the dashboard.