Moving towards showing events per ASN on dashboard
Replaced sources and targets graphs with a single "networks" graph.
Currently, the ASN is shown rather than the name but will fix this
shortly.

Extended size of event list to use some of the space made available
by removing the targets graph. This means long AS names will be
less likely to cause line breaks in the event group description.