Add options to add dashboard link and set project title
The dashboard link option allows us to enable or disable the "Dashboard"
link appearing at the top of the page, depending on whether a project is
using netevmon.

The project title simply allows us to change the test in the banner that
appears on each page to suit the project that our software is being used
for, e.g. BTM vs AMP vs other future projects.