Add an option to disable all event-related content
This includes events appearing on time series graphs and the entire
dashboard. This option is designed to be used when you do not have
a suitable event database available -- previously, we had to create a dummy
netevmon database for deployments that did not actually use netevmon just to
prevent ampy from constantly spouting errors about failed queries and
missing databases.