Merge event groups that share the exact same members
Calculate the necessary stats for the event time series and top
networks graphs when we parse the event groups for the dashboard. This
should save a fair bit of time fetching and re-processing the same
event groups multiple times (at the slight cost of making parse_event_groups
look hideous).

Try to make sure we don't count events or event groups multiple times when
the groups are merged.