Moved event parsing code into a separate class
This is mostly to re-factor the event parsing code so that it is not
a stupidly long function.

Removed the need for global site_count and event_timeseries variables. These
are now calculated whenever the events are parsed and then cached so that
our subsequent AJAX calls can just grab them from the cache.