Remove traceroute event groups that are subsets of other event groups
To be removed, a group must :
 * only contain traceroute events
 * only contain events that are present in the superset group
 * have the same start time as the superset group
 * have the same 'grouped by' ASNs/endpoints as the superset group

This should get rid of the "redundant" traceroute groups that result from
us creating a ?incr and a ?decr group for each traceroute event. Previously,
if we added an event to the ?incr group, the merging code would not recognise
that we no longer need the ?decr group (assuming no further events are added
to it later!).