Add 'split' to _matrix API parameters
This will allow us to adjust the matrix view link to match the currently
selected split on the matrix itself, i.e. if looking at an IPv4 only matrix,
clicking on a cell will take you to a graph with only an IPv4 line on it,
not both IPv4 and IPv6.