Shrink default detail graph size to 4 hours
This is to avoid the problem where the default aggregation was causing
the graph line to be misleading when compared with the value shown in
the matrix cell (when clicking through to a graph from a matrix cell).

The default detail graph size can now be overridden on a per-graphpage
basis: this has been used to keep the throughput default graph size at
48 hours as this test will generally be run much less frequently so
aggregation will not be a factor.