Implemented select2 dropdown for AS filtering on the dashboard
The AS names that are shown are loaded remotely using an AJAX call and are
paginated to prevent us from trying to load too many into the dropdown at once.

Also implemented the AJAX call to fetch the current filter parameters, so that
we will be able to update them and push them back to the database.

Started re-working the flow of how the events are loaded on the dashboard. Now
that we have more complicated filtering, our event processing becomes a bit
trickier and needs more thought. So far, this means that I've commented out the
code that actually fetches the events for now until I work out exactly what
I'm going to be doing.