Add file for tracking recent major changes to amp-web
This will hopefully serve as a useful pre-upgrade guide to the major changes
that will result from deploying a new version of amp-web and list all of the
associated upgrades that also need to occur at the same time (e.g. database
schema changes, upgrades of other software, etc.).

This file should also be used to track deployment dates of each version of a
on our various production systems (skeptic etc.) so that we have some proper
documentation of when rollouts happen and what versions have rolled out to
each system.

Bumped amp-web version number to 1.0.