Make event filtering interface dynamic
Event filtering properties are now loaded from the database via javascript
and are used to populate the event filter panel.

Changes to event filters using the panel are posted back to the server to
update the database, so they will persist between page reloads. Changes to
AS, source or target filter lists will cause those sections of the panel to be
reloaded to reflect the change in state.

Adding and removing AS-based filtering options now works, although no filters
are actually applied yet.

Toggle buttons for event types also change state when clicked on - red to show
event types that are removed, blue for event types that are included.