Make endpoint quantity filtering work and staggered event group fetching.
Endpoint quantity filtering means that the dashboard can be restricted to
only show event groups that affect at least a certain number of sources,
targets or endpoints in total.

Staggered event group fetching means that we'll only fetch 20 event groups
at a time, even if the user has asked for a large time range and maximum
number of event groups. If the user scrolls to the bottom of the event group
list, then we'll fetch the next batch of events (much like the old infinite
event list).

As a result, we now also allow a maximum of 0 groups which will keep fetching
groups until the start time limit is reached.

Removed the eventlist link from the dashboard as it is somewhat redundant now.