Event dashboard now updates with new or modified events
Note: this only happens if the filter endtime has not been modified.

New event groups will be inserted onto the dashboard and any event groups
that have changed are updated.

Only event groups that are relatively recent are going to be updated, so as
to avoid having to re-fetch event groups that probably won't be changing anyway.
Added ability to query only for event groups starting after a certain timestamp
to support this feature.

Current dashboard update frequency is 1 min, but this could be tweaked if
necessary.