Fix bug where guest filter names were persisting through logins
The problem was that the filter name from the cookie was being used, even
if that username and filtername combination didn't exist in the user filters
table. Now if the user is logged in and we encounter a filtername that doesn't
exist, we'll try again with "default" rather than automatically creating a
new filter with the missing name. Only if "default" fails will we create a
new filter (using "default" as the name), so we can still bootstrap a filter
for new users.