Fix bug where rapid filter changes were being ignored
Previously, we would ignore any new event fetch requests if there was an
existing request outstanding. This meant that multiple filtering changes
(e.g. clicking on multiple event type toggles in quick succession) would only
result in the first change having an effect on the event list.

Now, we abort any existing requests so we can make a new request with the
latest filtering information, so the event list that is shown matches up
with the filtering state displayed on the dashboard.