Make latency and loss graphs work for amp-udpstream
UDPStream is now an option when creating an amp-latency or amp-loss graph, alongside
the existing ICMP, TCPPing and DNS collections.

This also required us to change all graph modals to include a view style when
creating a new view (amp-udpstream is the first collection to support both amp-latency
and its own amp-udpstream view style, so we now have to explicitly specify this to
ampy).