Make sure the last datapoint is viewable on the graphs
The last datapoint used to have a x-width of 0. If the last datapoint
did not have a significant vertical component, it was practically invisible
on the graphs.

When using a non-trivial binsize, this made us look like we were further
behind than we really are -- especially important with Influx and its
delayed aggregation.

I've fixed this by adding a small 150 second spur for the last datapoint
on the smokeping and basic time series graphs, so that users have something
to mouse over which will give them a tooltip for that datapoint.

There may be other graph types which may have this problem, e.g. the
rainbow graph, but it was most egregious on the standard latency and loss
graphs so I've concentrated on those for now.