Fix crash when the dashboard tries to autoload after a cache flush
After flushing the cache, the dashboard can no longer access any info about
common events (as these are only populated when the dashboard is loaded).
This meant our cache lookup would return None, which we were treating as a
valid dictionary.

Now if the cache lookup fails, set common_events to be an empty dictionary.

TODO: maybe try and recalculate the common events if there is no valid cache
entry?