Fix (somewhat) issues with graph modals breaking after one use.
The new modal-loading behaviour in bootstrap apparently clears all of
the previous modal selections, so we have to force an update of the modal
contents when the modal is shown on subsequent occasions.

Unfortunately, this means that the modal won't remember the selections you
used when adding a series to an existing graph, so if you want to add another
similar series you'll have to start over from 'source', 'destination' etc.

I did experiment a bit with remembering the last set of selections and
trying to re-populate the modal inputs when reloading the modal, but for
whatever reason I couldn't get the values I was setting to stick and got
fed up with the whole thing :)