Make graph modals remember previous selections
Changes in bootstrap modal behaviour had meant that the previous modal
selections were not remembered after adding a new group to a view. This
used to happen and was quite handy when trying to add multiple 'similar'
groups to a view.

As a result, I've implemented this manually inside the modal javascript,
right down to remembering which aggregation method you chose.

As a side-effect, the horrible old 'createview' URL creation that used to
be implemented inside each modal's submit() function has been replaced
with the submitAjax() function. The big advantage of this is that we can
make changes to the createview URL format (e.g. add a viewstyle!) and only
have to update the top-level modal.js.