Add a findMinimumY function to the basicTS graphstyle
A non-standard graph style may not work so well with an autoscaling y-axis
if it is not obvious as to which values in the series are the y values.
This change means that these graph styles can provide their own method for
finding the minimum Y rather than relying on the plotting library that won't
be able to understand their data.