Add ability for users to rate displayed events
User feedback is then saved into a file on the server's disk, which the
developers can then fetch and peruse at some later date.

Events can be rated by either clicking on the event marker on a graph view or
by clicking the rate button alongside the event on the dashboard.

Also tweaked the event links on the dashboard to always open the eventview
in a new tab.