Only re-fetch event groups after any filter changes are complete
Previously we were starting the fetch at the same time as the POST that
was being used to update the filters. On a relatively unloaded server, this
was fine -- the filter would be changed by the time the event fetcher wanted
to use it -- but on a more loaded server, we would end up with situations where
changes to filters were not being committed in time for them to be applied
properly.