Extended event group cleansing to remove redundant ASN groupings
For instance, consider a set of chronologically similar event groups: AS2,
AS1 + AS2, and AS1 + AS2 + AS3. Each event group contains a different set of
events, but this is simply an artefact of the paths required to reach the
their targets. It's easy for the human eye to realise that AS2 is almost
certainly the AS that is causing these events, because it is common to all
three event groups.

This commit adds extra checks to the cleansing code to check for situations
like this, i.e. groups having common ASNs, similar start times and
non-overlapping events, and combining the groups together with only the common
ASN(s) being retained in group label.