Added missing summary graph to traceroute path map view.
The graph itself is a simple line graph of the number of unique paths seen
during that bin. This is enough for users to see when there might be some
abnormal variety, as well as to see where there are gaps in the data.

The main benefit, though, is that the selection handles are now always
visible on the summary graph so it is a lot easier to change the time range
covered by the detail graph and see which time periods you are actually
looking at!