Replace cleansing function with a simpler cut-down version
Netevmon is going to handle most of the subset/superset detection and
removal now, so we only need to do the most obvious group merging
inside amp-web. In particular, netevmon won't look at recent groups
so we still have to careful to merge the particularly egregarious
double-ups before we put them on the dashboard.