Make sure we remove any cleansed events from the dashboard.
Event groups are displayed on the dashboard as soon as they are
created. If an event group then subsequently gets cleansed by netevmon,
we need to remove the panel for it from the dashboard, otherwise we'll
get event groups appearing out of chronological order.

The solution is to remember which panels we have on the dashboard and
make sure that any panels covering the time period that we've just queried
for are removed before we process the incoming event groups. We also need
to remember if a panel was open or closed, so that we don't start closing
an event group that someone was looking at just because we hit a minute
boundary.