Fix bug where groups could appear multiple times on the dashboard
There was a rare case where a group could be added to the dashboard
but not removed from the set of merging candidates, so it would then
be added again later on.

The problem occurs if a group has the same ASN set as a later group,
so we were confusing the two groups when trying to purge the expired
group.