Prevent multiple panels for the same group appearing on the dashboard.
It appears that the bug in the eventparser that means a group gets added
to the group list multiple times may not be entirely fixed. Until I can
create an easily replicable test case to track it down properly, this
code will at least try to prevent the group from appearing multiple
times on the dashboard.