Added a matrix for HTTP page sizes
I originally implemented this to test that my instructions for adding new
metrics were correct, but the end result has turned out interesting enough
that I've committed it back.

Note that for the moment, this results in a new matrix tab -- in the long run,
I'd like to combine the two HTTP matrix tabs into a single tab, but that'll
require a bit more work than I can put in right now.