Escape site/mesh names so that they match the correct routes.
WSGI will unquote the URI and then use that to try to match routes, so any sites that include slashes in the name will not match the correct route. Instead we have to double-escape those characters so that WSGI only unquotes one layer of quoting and then our code unquotes the second layer.