Don't colour matrix cells red for minor latency changes
If the stddev is very low, even minor shifts can appear alarming
on the matrix if we only use stddev as the basis for colouring
relative latency.

Any latency changes < 1ms are always coloured green.
Any latency changes < 5ms are coloured light-green.
All other changes are coloured as before (based on stddev).

Thresholds are not set in stone -- feel free to tweak if needed.