Fix slow dashboard loading, disable dashboard graphs.
Our dashboard pre-processing was taking too long, mainly
because we would call out to ampy for each event group
seen in the last 24 hours.

We only need to do this to generate data for the dashboard
graphs (most common events, events over time, etc.). For now
I've disabled these graphs globally so that we can concentrate
on getting the dashboard on screen in a timely manner.

In the long-term, these graphs should be loaded via an ajax call
and their processed data should be cached.

Also re-jigged the event parsing process in eventparser.py, so
we avoid doing unnecessary work on event groups that have no
chance of appearing on the dashboard right now anyway.