Duplicate yaml schedule endpoint to allow finer grained control.
Previously the schedules were available to anyone. Now the schedule
is always available to a logged in website user with permission to
view configuration (using a new route), which allows for the existing
route to be locked down through apache directives if desired (e.g.
require client to present an SSL certificate to get their config).